small progress toward working config
