+++ commands/generate/pages/edit/generateEdit.ts
@@ -1,4 +1,9 @@
+// import { readNextConfig } from "../../../../utils"
+
 export const generateEdit = (modelName: string, upperCaseFirstLetterModelName: string, finalJsonBodyItems: string, finalEditFormFieldItems: string) => {
+
+  // Checks next config for issues and alerts user accordingly
+  // const { database, style } = readNextConfig()
 
   const editPage = `
   import { useRouter } from "next/router";
