+++ utils.js
@@ -47,7 +47,7 @@ const readNextConfig = () => {
     }
     else {
         console.log("No nextGenConfig.json found");
-        return false;
+        return JSON.parse("false");
     }
 };
 exports.readNextConfig = readNextConfig;
